requirements.txt
Django>=6.0
psycopg[binary]>=3.1
python-dotenv>=1.0.0
djangorestframework>=3.15
django-cors-headers>=4.0
